+++ src/components/account.js
@@ -4,7 +4,7 @@ const Account = () => {
     return (
         
         <div>
-            <h1>This page will show the users personal information and allow them to update it.</h1>
+            <h1>This is the User Account page. It will show the users personal information and allow them to update it.</h1>
         </div>
         
     );
