+++ src/components/dashboard.js
@@ -4,7 +4,7 @@ const Dashboard = () => {
     return (
         
         <div>
-            <h1>This page will show different content based on the user type.</h1>
+            <h1>This is the User Dashboard. It will show different content based on the user type.</h1>
         </div>
         
     );
